# openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SmartUnderwrite.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/Applications": {
      "post": {
        "tags": [
          "Applications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AffiliateId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "CreatedAfter",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "CreatedBefore",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "MinAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MaxAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "ProductType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfLoanApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfLoanApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfLoanApplicationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{id}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplicationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{id}/evaluate": {
      "post": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{id}/decision": {
      "post": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{id}/documents": {
      "post": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "ContentType": {
                        "type": "string"
                      },
                      "ContentDisposition": {
                        "type": "string"
                      },
                      "Headers": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "Length": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "Name": {
                        "type": "string"
                      },
                      "FileName": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "description": {
                        "type": "string",
                        "default": null
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{applicationId}/documents": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentUploadResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentUploadResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentUploadResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Applications/{applicationId}/documents/{documentId}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Applications/{applicationId}/documents/{documentId}/download": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateApplicationRequest": {
        "required": [
          "firstName",
          "lastName",
          "ssn",
          "dateOfBirth",
          "phone",
          "email",
          "address",
          "productType",
          "amount",
          "incomeMonthly",
          "employmentType"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "ssn": {
            "maxLength": 11,
            "minLength": 9,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "productType": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "amount": {
            "maximum": 1000000,
            "minimum": 1000,
            "type": "number",
            "format": "double"
          },
          "incomeMonthly": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "employmentType": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "creditScore": {
            "maximum": 850,
            "minimum": 300,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
      "LoanApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "affiliateId": {
            "type": "integer",
            "format": "int32"
          },
          "affiliateName": {
            "type": "string"
          },
          "applicant": {
            "$ref": "#/components/schemas/ApplicantDto"
          },
          "productType": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "incomeMonthly": {
            "type": "number",
            "format": "double"
          },
          "employmentType": {
            "type": "string"
          },
          "creditScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            }
          },
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionDto"
            }
          }
        }
      },
...
      "ManualDecisionRequest": {
        "required": [
          "outcome",
          "reasons"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "reasons": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "justification": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          }
        }
      },
      "PagedResultOfLoanApplicationDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanApplicationDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}